requirements.txt
anyio==3.6.2
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==1.16.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.75.0
h11==0.14.0
httptools==0.2.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
orjson==3.8.10
packaging==23.0
passlib==1.7.4
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pydantic==1.9.0
pymongo==3.11.0
pytest==7.0.1
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.17.1
tomli==2.0.1
typing_extensions==4.5.0
ujson==4.3.0
urllib3==1.26.15
uvicorn==0.15.0
uvloop==0.17.0
watchgod==0.8.2
websockets==11.0.1
